Refactored code; connected DB to remote cluster;
Refactored code into routes and controller for cleaner structure and
modularity. All routes go to the routes file and each method calls
a function from the controller, which details the specific req and res.
The connection of mongoose has been modified to connect to a remote DB
from mogoDB cloud. Connection seems to be working but the model schemas
are seemingly not recorded.