XMLHttprequest for Login Page
Using axios, we are now able to retrieve the repsone from backend
after posting our request to it.
The axios would return the following 3 variable:
response => all the response, including data, config, request, etc.

resData => the data field of the response

resStatus => Status field that returns whether we login successfully
or not in the scope of the server side

resCookie => the cookie returned from the server side for this login