Check Password Breached feature: fetch cookie resolved
localStorgae is implied to fetch cookie for curren user's session.
this will not compile in this branch for functionality from web-dev
branch has not merged with ui-ux at this point yet.

if anyone who wants to run this feature without incorporating
the web-dev functionalities, change value in line 18 with an
active cookie. (example given in line 19)